+++ AeonCore/IReadOnlyStats.cs
@@ -9,7 +9,7 @@
 		public int RawValue<TStat>() where TStat : StatType, new() => 
 			Get<TStat>().Value;
 
-		public double Converted<TStat>() where TStat : StatType, new() => 
+		public decimal Converted<TStat>() where TStat : StatType, new() => 
 			Get<TStat>().Converted;
 
 		public int ConvInt<TStat>() where TStat : StatType, new() => 
@@ -21,7 +21,7 @@
 		public int DynamicValue<TStat>() where TStat : StatTypeDynamic, new() =>
 			GetDyn<TStat>().Value;
 
-		public double DynConverted<TStat>() where TStat : StatTypeDynamic, new() =>
+		public decimal DynConverted<TStat>() where TStat : StatTypeDynamic, new() =>
 			GetDyn<TStat>().Converted;
 
 		public int DynConvInt<TStat>() where TStat : StatTypeDynamic, new() =>
